♻️ (sample): Create a compoment for the check dapp drawer
