+++ apps/sample/src/components/CalView/index.tsx
@@ -1,25 +1,11 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import {
-  Button,
-  Divider,
-  Flex,
-  Grid,
-  Icons,
-  InfiniteLoader,
-} from "@ledgerhq/react-ui";
+import React, { useCallback, useState } from "react";
+import { Grid } from "@ledgerhq/react-ui";
 
-import { Block } from "@/components/Block";
 import { ClickableListItem } from "@/components/ClickableListItem";
-import {
-  CommandForm,
-  ValueSelector,
-} from "@/components/CommandsView/CommandForm";
 import { PageWithHeader } from "@/components/PageWithHeader";
 import { StyledDrawer } from "@/components/StyledDrawer";
-import { FieldType } from "@/hooks/useForm";
 
-import { CalAvailabilityResponseComponent } from "./CalAvailabilityResponse";
-import { checkContractAvailability, Descriptor } from "./CalNetworkDataSource";
+import { CalCheckDappDrawer } from "./CalCheckDappDrawer";
 
 const CAL_SERVICE_ENTRIES = [
   {
@@ -28,185 +14,42 @@ const CAL_SERVICE_ENTRIES = [
   },
 ];
 
-export type CalActionProps<
-  _,
-  Input extends Record<string, FieldType> | void,
-> = {
-  title: string;
-  description: string;
-  initialValues: Input;
-  validateValues?: (args: Input) => boolean;
-  valueSelector?: ValueSelector<FieldType>;
-};
-
-type Response = {
-  searchAddress: string;
-  date: Date;
-  responseType: string;
-  result: Descriptor[]; // Store the result from the API
-  loading: false;
-  id: number;
-};
-
-export function CalActionDrawer<
-  Output,
-  Input extends Record<string, FieldType>,
->(props: CalActionProps<Output, Input>) {
-  const { initialValues, valueSelector, validateValues } = props;
-
-  const nonce = useRef(-1);
-  const [values, setValues] = useState<Input>(initialValues);
-  const [valuesInvalid, setValuesInvalid] = useState<boolean>(false);
-  const [responses, setResponses] = useState<Response[]>([]);
-  const [loading, setLoading] = useState(false);
-  const handleClickExecute = useCallback(() => {
-    setLoading(true);
-    const id = ++nonce.current;
-
-    const fetchData = async () => {
-      try {
-        console.log("Trigger Request to checkContractAvailability");
-
-        const response = await checkContractAvailability(
-          values.smartContractAddress.toString(),
-          values.calUrl.toString(),
-          values.branch.toString(),
-        );
-
-        setResponses((prev) => [
-          ...prev,
-          {
-            searchAddress: values.smartContractAddress.toString(),
-            date: new Date(),
-            responseType: response.responseType,
-            result: response.descriptors, // Store the result from the API
-            loading: false,
-            id,
-          },
-        ]);
-
-        setLoading(false);
-      } catch (_error) {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, [values]);
-
-  const handleClickClear = useCallback(() => {
-    setResponses([]);
-  }, []);
-
-  const responseBoxRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (responseBoxRef.current) {
-      responseBoxRef.current.scrollTop = responseBoxRef.current.scrollHeight;
-    }
-  }, [responses]);
-
-  useEffect(() => {
-    if (validateValues) {
-      setValuesInvalid(!validateValues(values));
-    }
-  }, [validateValues, values]);
-
-  return (
-    <>
-      <Block>
-        <Flex
-          flexDirection="column"
-          opacity={loading ? 0.5 : 1}
-          rowGap={3}
-          pointerEvents={loading ? "none" : "auto"}
-        >
-          <CommandForm
-            initialValues={values}
-            onChange={setValues}
-            valueSelector={valueSelector}
-            disabled={loading}
-          />
-          <Divider />
-        </Flex>
-        <Flex flexDirection="row" flex={1} columnGap={3}>
-          <Button
-            variant="main"
-            onClick={handleClickExecute}
-            disabled={loading || valuesInvalid}
-            Icon={() =>
-              loading ? <InfiniteLoader size={20} /> : <Icons.ArrowRight />
-            }
-          >
-            Check Availability
-          </Button>
-        </Flex>
-      </Block>
-      <Block flex={1} overflowY="hidden">
-        <Flex
-          ref={responseBoxRef}
-          flexDirection="column"
-          rowGap={4}
-          flex={1}
-          overflowY="scroll"
-        >
-          {responses.slice().map((response, key) => (
-            <CalAvailabilityResponseComponent
-              key={key}
-              type={response.responseType}
-              date={response.date}
-              loading={response.loading}
-              descriptors={response.result}
-              searchAddress={response.searchAddress}
-            />
-          ))}
-        </Flex>
-        <Button
-          variant="main"
-          outline
-          onClick={handleClickClear}
-          disabled={responses.length === 0}
-        >
-          Clear responses
-        </Button>
-      </Block>
-    </>
-  );
-}
-
 export const CalView = () => {
-  const [isOpen, setIsOpen] = useState(false);
-  const openDrawer = useCallback(() => {
-    setIsOpen(true);
+  const [isCheckDappOpen, setIsCheckDappOpen] = useState(false);
+  const openCheckDapp = useCallback(() => {
+    setIsCheckDappOpen(true);
   }, []);
 
-  const closeDrawer = useCallback(() => {
-    setIsOpen(false);
+  const closeCheckDapp = useCallback(() => {
+    setIsCheckDappOpen(false);
   }, []);
 
-  const title = "Check dApp availability";
-  const description = "Check descriptor availability on the CAL";
+  const pageTitle = "Check dApp availability";
+  const pageDescription = "Check descriptor availability on the CAL";
 
   return (
     <PageWithHeader title="Crypto Assets">
-      <Grid columns={1}>
+      <Grid
+        columns={1}
+        style={{ rowGap: 6, columnGap: 6, overflowY: "scroll" }}
+      >
         {CAL_SERVICE_ENTRIES.map(({ title, description }) => (
           <ClickableListItem
             key={`keyring-${title}`}
             title={title}
             description={description}
-            onClick={openDrawer}
+            onClick={openCheckDapp}
           />
         ))}
       </Grid>
       <StyledDrawer
-        isOpen={isOpen}
-        onClose={closeDrawer}
+        isOpen={isCheckDappOpen}
+        onClose={closeCheckDapp}
         big
-        title={title}
-        description={description}
+        title={pageTitle}
+        description={pageDescription}
       >
-        <CalActionDrawer
+        <CalCheckDappDrawer
           title={""}
           description={""}
           initialValues={{
